Fix OpticalTrain shared cmds attribute and fix docstring
When instantiating OpticalTrain with a UserCommands object (e.g. loading instrument from IRDB), make a deepcopy of the argument to avoid multiple instances of OpticalTrain using the same cmds object in memory. This avoids that changing a cmds parameter in a specific system changes all other systems instantiated with the same UserCommands object.
Additionally, simplify logic and improve docstring to reflect the possibility of using a str argument in load().