Raise NotImplementedError instead of NotImplemented
Justification (from "Flake8Rules"):
NotImplemented is a special value which should be returned by the binary
special methods to indicate that the operation is not implemented with
respect to the other type. Raise NotImplementedError to indicate that a
super-class method is not implemented and that child classes should
implement it.