Avoid in-place operations because of dtype conflicts
This caused issues in some cases due to the dtype change that can occur in the Quantization effect. The error was seen with the DarkCurrent class, but could potentially also happen elsewhere, so I defused all potential cases here. If this is really impacting performance or memory usage, we should try to find a better solution, e.g. converting the dtype of the to-be-added array before the in-place operation.