Deepcopy readout, to prevent it being overwritten. Closes #439
Create a copy of the readouts, because they are still referenced here.
That is, self.detector_managers[0][0].hdu is hdul[1].
Subsequent readouts would therefor overwrite earlier readouts
without the copy.
It is necessary to keep the detector readouts in memory, because
this makes it possible to add support for up-the-ramp sampling,
that is, subsequent readouts without reseting the detector inbetween.