Use common retriever for list_example_data
This function used to scrape the example_data directory on the server to
find all available files there. However, as we are now using pooch to
perform the download of these files, which only allows downloading ones
with a known hash in the registry file, this list of available files is
by definition equal to the list of files in the registry. It is thus no
longer required to query the server for that information. Also, if a file
gets added to the server, but not (yet) to the registry, this function
would have produced a wrong result, in showing a file that is not actually
downloadable with download_example_data.