Revive forgotten test
This test was indented, thus an internal function of another test and as
such not discovered by pytest. There seems to be nothing wrong with it
though, apart from some old fixtures that are now functions (not sure
which is better but that's what we have now) and a missing import.