Improve handling of Source.meta and fields.meta
Some code (mostly in ScopeSim_Templates) relies on single-field sources
exposing their field meta as the Source's meta. If there's more than one
field, the metas should always be accessed through the fields to avoid
ambiguity. This (hacky) solution now seems to work and is certainly better
than silently always returning just the first field's meta.

Also removed the setter for meta. It's a dict, so if it needs to be updated,
just do that. It should never be replaced (== overwritten!) anyway.