Make z_order a class attribute
I know, we want to get rid of z_order eventually. But that's for later.
Now, in the context of the effects dataclasses change, and the changes
to meta associated with that, I need to make this move anyway. I realized
this can be done independently, so why not. Also, conceptually, the
z_order is a property of the class, not the instance. As in, any instance
of an effect class will not change the z_order. That is, except
AtmosphericDispersionCorrection apparently, but that should work as well.