Generatorify stringify_dict function
This aux function is only used in HDU.header.update() calls, which can eat
pair-generators just fine (yes I tested that). This also removed the need
for creating a deepcopy of the original dict in memory (not that it
matters much but still), because the input dict is never mutated anyway.
Looking at the internal implementation of HDU.header.update(), this
actually does mean in this case that the generator is processed
item-by-item and never turned into a full dict or anything.

In theory, this could be expanded in the future to also process comments,
e.g. whenever the dict value is a 2-tuple, simply yield a 3-tuple, which
HDU.header.update() will understand as a key-value-comment group...