Raise instead of log if package path not found
I found out the creating a UserCommands object from an existing package
and then another one from a non-existing one (i.e. not installed or
misspelled) will log a small warning, but then leave the previous package's
settings in the cmds! This could potentially lead to hard-to-track bugs
and I don't see why we shouldn't just raise in this case.

For the case of a sub-package not found (e.g. ELT for METIS), we can still
proceed, this might be fine in some cases (e.g. METIS WCU modes).