Store fields in FOV.xy_fields lists as SourceFields
I.e., turn around FOV.fields and FOV.xy_fields. We need the individual
ones much more often than the total one and this makes a few things
easier and more efficient.

Importantly, this means the order of individual fields from the source is
no longer preserved when extracted by the FOV. This shouldn't matter in
terms of results, but some of our tests relied of that, so they needed to
be adaped accordingly.