Make better use of `SourceField` in `FieldOfView`
Refactor most of those `FieldOfView._make_xy()` methods to use the
`SourceField.spectra/um` attribute directly, effectively eliminating the
need for the `"SPEC_REF"` header keyword.

This also removed the need for a separate `FieldOfView.spectra` attribute.
I kept it for now as a property since it's still used in a few other
places I think.