Actually use the power of SourceFields in the FOV
Refactor most of those _make_xy methods to use the SourceField.spectra/um
attribute directly, effectively eliminating the need for SPEC_REF.

This also removed the need for a separate FOV.spectra attribute. I kept it
for now as a property since it's still used in a few other places.