Fix output cube BUNIT
The keyword was originally set _before_ some lines that actually modify it,
and never updated. There are some conflicts in various docstrings and
comments, mostly centered around the "(c)m-2" (TEL area), but this should
now be the correct unit (and the one that the spectral traces have been
assuming anyway).