Bring dependency versions back in sync with lock file
These got out of sync because of dependabot. It's not critical, but our
minimumdependency approach mandates that the versions in the lock file
should be equal to the minimum specified version in the toml file.

Fortunatly, this now doesn't include any of our "important" dependencies,
so I don't expect any major surprises here (famous last words).