Add proper error message for removed `Quantization` effect
When running ScopeSim v>=0.10.0 with an old METIS version, the error was just `AttributeError: module 'scopesim.effects' has no attribute 'Quantization'`, rather unhelpful to the end user.
Now it shows what the actual "problem" is.
Given that ScopeSim and the IRDB packages are released separately, it could be argued that the `Quantization` effect should have been deprecated more carefully.
But then again, it could also be argued that it shouldn't have existed in the first place.
And we're still in a 0.x version after all, so whatever.
This is better then nothing.