Refactor `OpticalElement.get_all()`
This used to call `effect_utils.get_all_effects()`, which was basically the
same as `OpticalElement._get_matching_effects()`, except that it returns
only included effect. Some other methods seem to relied on
`OpticalElement._get_matching_effects()` returning _all_ effects, so I
added a flag to that methods that defaults to this, but is changed by the
call in `OpticalElement.get_all()`. This ensures that no behavior changes
to the outside.

FWIW, `effect_utils.get_all_effects()` seems to have been written in a much
too complicated way, not considering that `isinstance()` can deal with
multiple classes at the same time out of the box...