Update dependabot.yml
Apparently there's a maximum cooldown of 90 days. Which is fine, but I specifically looked in the documentation before setting this value, to see if there indeed is such a maximum value. Looked twice in fact, and couldn't find any mention of a maximum. So either I'm blind as a bat, or GitHub's documentation is shite. Hmmm

The property '#/updates/2/cooldown/semver-major-days' did not have a maximum value of 90, inclusively